Implement prototype
This prototype implements the base game without scoring. It has some
quality-of-life features like DAS and a grid. It lacks a scoring system
and the "random bag" selection for pieces.

There is much work to do in terms of visuals. Block colors can be
customized by listing their hex codes in a colors.txt file in order, but
there is no other customization. The game renders in the top-left of an
otherwise-empty white screen.