Update rendering and move data to structs
This commit reorganizes main.c by moving most of the game data into
structs. The file is still a bit messy, but having this modularity will
let us break it up in the future.

Holding has been implemented, the next pieces are rendered on the side
of the screen, and the board now appears centered in the window.