Add support for background shaders
A "sky" shader has been added which is displayed in the background of
the game board. This commit experiments with a number of game parameters
exposed to it, including the board height and time since game over. It
is in an incomplete state, as I would like to break up the code in
main.c before continuing with visuals.